fix: adjust hook details
