+++ src/hooks/moneyDestination/useDetailsDestination/index.ts
@@ -1,0 +1,28 @@
+import { useState } from "react";
+import { useEvaluateRuleByBusinessUnit } from "@hooks/rules/useEvaluateRuleByBusinessUnit";
+import { IEntry } from "@components/data/Table/types";
+import { IAppData } from "@ptypes/context/authAndPortalDataProvider/IAppData";
+
+const useDetailsDestination = (appData: IAppData, data: IEntry) => {
+  const [showModal, setShowModal] = useState(false);
+
+  const handleToggleModal = () => {
+    setShowModal(!showModal);
+  };
+
+  const { evaluateRuleData } = useEvaluateRuleByBusinessUnit(
+    appData.businessUnit.publicCode,
+    {
+      ruleName: "LineOfCredit",
+      conditions: [
+        {
+          condition: "MoneyDestination",
+          value: data.descriptionUse,
+        },
+      ],
+    },
+  );
+
+  return { showModal, handleToggleModal, evaluateRuleData };
+};
+export { useDetailsDestination };
