+++ src/pages/moneyDestination/tabs/moneyDestinationTab/tools/details/index.tsx
@@ -1,9 +1,9 @@
-import { useContext, useState } from "react";
+import { useContext } from "react";
 
 import { IEntry } from "@components/data/Table/types";
-import { useEvaluateRuleByBusinessUnit } from "@hooks/rules/useEvaluateRuleByBusinessUnit";
 import { AuthAndPortalData } from "@context/authAndPortalDataProvider";
 import { DetailsMoneyDestination } from "@design/feedback/detailsMoneyDestination";
+import { useDetailsDestination } from "@src/hooks/moneyDestination/useDetailsDestination";
 
 interface IDetails {
   data: IEntry;
@@ -11,25 +11,10 @@ interface IDetails {
 
 const Details = (props: IDetails) => {
   const { data } = props;
-  const [showModal, setShowModal] = useState(false);
   const { appData } = useContext(AuthAndPortalData);
 
-  const handleToggleModal = () => {
-    setShowModal(!showModal);
-  };
-
-  const { evaluateRuleData } = useEvaluateRuleByBusinessUnit(
-    appData.businessUnit.publicCode,
-    {
-      ruleName: "LineOfCredit",
-      conditions: [
-        {
-          condition: "MoneyDestination",
-          value: data.descriptionUse,
-        },
-      ],
-    },
-  );
+  const { showModal, handleToggleModal, evaluateRuleData } =
+    useDetailsDestination(appData, data);
 
   return (
     <>
